requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
eventlet==0.36.1
anthropic==0.28.0
python-dotenv==1.0.1
requests==2.32.3
colorama==0.4.6
rich>=13.0.0          # SentinelAI dashboard (terminal) + logging
cryptography==42.0.8

# attack_sim / network-layer integration (no new deps required —
# attack_sim uses only Python stdlib: socket, threading, urllib, json, os)
# Optional: groq for LLM-enhanced detection patches (already present via groq client)
